+++ testapps/GH991/src/main/java/test/AppEntry.java
@@ -1,6 +1,8 @@
 package test;
 
 import act.Act;
+import act.event.OnEvent;
+import act.util.LogSupport;
 import act.ws.*;
 import org.osgl.mvc.annotation.GetAction;
 
@@ -12,15 +14,26 @@ import org.osgl.mvc.annotation.GetAction;
  * in the browser!
  */
 @SuppressWarnings("unused")
-public class AppEntry {
+@WsEndpoint("/ws2")
+public class AppEntry extends LogSupport {
 
     @WsEndpoint("/ws")
-    public static class WsConnHandler implements WebSocketConnectionListener {
+    public static class WsConnHandler extends LogSupport implements WebSocketConnectionListener {
         @Override
         public void onConnect(WebSocketContext context) {
             int n = context.manager().urlRegistry().count();
             context.send("count: " + n);
         }
+
+        @Override
+        public void onClose(WebSocketContext context) {
+            info("Connection closed: " + context.url());
+        }
+    }
+
+    @OnEvent
+    public void onClose(WebSocketCloseEvent event) {
+        info(">>> Connection closed: " + event.source().url());
     }
 
     @GetAction